+++ src/main/java/com/farmtastic/farm/controller/SensorDataController.java
@@ -1,6 +1,7 @@
 package com.farmtastic.farm.controller;
 
-import com.farmtastic.farm.dto.ZoneDashboardResponse;
+import com.farmtastic.farm.dto.ZoneHistoryResponse;
+import com.farmtastic.farm.dto.ZoneLatestResponse;
 import com.farmtastic.farm.service.SensorDataService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
@@ -19,16 +20,16 @@ public class SensorDataController {
     // 최신 상태 조회 API
     // GET /api/zones/{zoneId}/latest
     @GetMapping("/{zoneId}/latest")
-    public ResponseEntity<ZoneDashboardResponse> getLatestSensorData(@PathVariable("zoneId") Long zoneId) {
-        ZoneDashboardResponse response = sensorDataService.getLatestData(zoneId);
+    public ResponseEntity<ZoneLatestResponse> getLatestSensorData(@PathVariable("zoneId") Long zoneId) {
+        ZoneLatestResponse response = sensorDataService.getLatestData(zoneId);
         return ResponseEntity.ok(response);
     }
 
     // 이력 조회 API
     // GET /api/zones/{zoneId}/history
     @GetMapping("/{zoneId}/history")
-    public ResponseEntity<ZoneDashboardResponse> getSensorDataHistory(@PathVariable("zoneId") Long zoneId) {
-        ZoneDashboardResponse response = sensorDataService.getHistoryData(zoneId);
+    public ResponseEntity<ZoneHistoryResponse> getSensorDataHistory(@PathVariable("zoneId") Long zoneId) {
+        ZoneHistoryResponse response = sensorDataService.getHistoryData(zoneId);
         return ResponseEntity.ok(response);
     }
 }
