+++ src/main/java/com/farmtastic/farm/service/SensorDataService.java
@@ -4,7 +4,8 @@ import com.farmtastic.farm.domain.SensorLog;
 import com.farmtastic.farm.domain.Zone;
 import com.farmtastic.farm.domain.enums.DeviceType;
 import com.farmtastic.farm.domain.enums.ModelType;
-import com.farmtastic.farm.dto.ZoneDashboardResponse;
+import com.farmtastic.farm.dto.ZoneHistoryResponse;
+import com.farmtastic.farm.dto.ZoneLatestResponse;
 import com.farmtastic.farm.repository.SensorLogRepository;
 import com.farmtastic.farm.repository.ZoneRepository;
 import jakarta.persistence.EntityNotFoundException;
@@ -26,33 +27,32 @@ public class SensorDataService {
     private final ZoneRepository zoneRepository;
 
     // 특정 구역의 최신 센서 데이터를 조회
-    public ZoneDashboardResponse getLatestData(Long zoneId) {
+    public ZoneLatestResponse getLatestData(Long zoneId) {
         Zone zone = zoneRepository.findById(zoneId)
                 .orElseThrow(() -> new EntityNotFoundException("Zone not found with id: " + zoneId));
 
         List<SensorLog> latestLogs = sensorLogRepository.findLatestLogs(zoneId);
 
         // ModelType 별 최신 값을 Map으로 구성
-        Map<ModelType, ZoneDashboardResponse.LatestValue> latestValuesMap = latestLogs.stream()
+        Map<ModelType, ZoneLatestResponse.LatestValue> latestValuesMap = latestLogs.stream()
                 .collect(Collectors.toMap(
                         log -> log.getDevice().getModelType(),
-                        log -> ZoneDashboardResponse.LatestValue.builder()
+                        log -> ZoneLatestResponse.LatestValue.builder()
                                 .value(log.getLogValue().doubleValue())
                                 .timestamp(log.getLogTime())
                                 .build(),
                         (existing, replacement) -> existing // 동일 키 충돌 시 기존 값 유지
                 ));
 
-        return ZoneDashboardResponse.builder()
+        return ZoneLatestResponse.builder()
                 .zoneId(zone.getZoneId())
                 .zoneName(zone.getZoneName())
                 .latestValues(latestValuesMap)
-                .historyValues(null)
                 .build();
     }
 
     // 특정 구역의 최근 3시간 센서 데이터 이력을 조회
-    public ZoneDashboardResponse getHistoryData(Long zoneId) {
+    public ZoneHistoryResponse getHistoryData(Long zoneId) {
         Zone zone = zoneRepository.findById(zoneId)
                 .orElseThrow(() -> new EntityNotFoundException("Zone not found with id: " + zoneId));
 
@@ -62,11 +62,11 @@ public class SensorDataService {
         List<SensorLog> recentLogs = sensorLogRepository.findLogsAfter(zoneId, DeviceType.SENSOR, startTime);
 
         // 가져온 로그들을 ModelType을 기준으로 그룹화
-        Map<ModelType, List<ZoneDashboardResponse.HistoryPoint>> historyValuesMap = recentLogs.stream()
+        Map<ModelType, List<ZoneHistoryResponse.HistoryPoint>> historyValuesMap = recentLogs.stream()
                 .collect(Collectors.groupingBy(
                         log -> log.getDevice().getModelType(),
                         Collectors.mapping(
-                                log -> ZoneDashboardResponse.HistoryPoint.builder()
+                                log -> ZoneHistoryResponse.HistoryPoint.builder()
                                         .timestamp(log.getLogTime())
                                         .value(log.getLogValue().doubleValue())
                                         .build(),
@@ -74,7 +74,7 @@ public class SensorDataService {
                         )
                 ));
 
-        return ZoneDashboardResponse.builder()
+        return ZoneHistoryResponse.builder()
                 .zoneId(zone.getZoneId())
                 .zoneName(zone.getZoneName())
                 .historyValues(historyValuesMap)
